refactor: workspace ui & form
