--- crates/frontend/src/components/input_components.rs
@@ -1,90 +1,0 @@
-use leptos::*;
-use serde_json::{Map, Value};
-
-use crate::components::dropdown::{Dropdown, DropdownBtnType, DropdownDirection};
-
-#[component]
-pub fn enum_dropdown(
-    schema: Map<String, Value>,
-    config_value: String,
-    handle_change: Callback<String, ()>,
-    #[prop(default = String::new())] class: String,
-    #[prop(default = false)] disabled: bool,
-    #[prop(default = "")] name: &'static str,
-) -> impl IntoView {
-    let (value, set_value) = create_signal(config_value.replace("\\", ""));
-    let (selected_enum, set_selected_enum) = create_signal(String::from("Choose Enum"));
-
-    let enum_array: Vec<String> = schema
-        .get("enum")
-        .and_then(|v| v.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    create_effect({
-        let value = value.clone();
-        let enum_array = enum_array.clone();
-        move |_| {
-            if !value.get().is_empty() && enum_array.contains(&value.get()) {
-                set_selected_enum.set(value.get());
-            }
-        }
-    });
-
-    view! {
-        <div class=format!(
-            "form-control {}",
-            class,
-        )>
-            {move || {
-                view! {
-                    <Dropdown
-                        disabled=disabled
-                        dropdown_width="w-100"
-                        dropdown_icon="".to_string()
-                        dropdown_text=(move || { selected_enum.get() })()
-                        dropdown_direction=DropdownDirection::Down
-                        dropdown_btn_type=DropdownBtnType::Select
-                        dropdown_options=enum_array.clone()
-                        name
-                        on_select=Callback::new(move |selected: String| {
-                            handle_change.call(selected.clone());
-                            set_value.set(selected.clone());
-                            set_selected_enum.set(selected.clone());
-                        })
-                    />
-                }
-            }}
-
-        </div>
-    }
-}
-
-#[component]
-pub fn boolean_toggle(
-    value: bool,
-    on_change: Callback<bool, ()>,
-    #[prop(default = String::new())] class: String,
-    #[prop(default = false)] disabled: bool,
-    #[prop(default = "")] name: &'static str,
-) -> impl IntoView {
-    let (flag, set_flag) = create_signal(value);
-    view! {
-        <input
-            disabled=disabled
-            on:click=move |_| {
-                set_flag.update(|val| *val = !*val);
-                on_change.call(flag.get());
-            }
-
-            type="checkbox"
-            name=name
-            class=format!("toggle toggle-[#ffffff] flex items-center {class}")
-            checked=flag
-        />
-    }
-}
