+++ crates/superposition/src/workspace/types.rs
@@ -1,19 +1,4 @@
 use serde::Deserialize;
-use superposition_types::database::models::WorkspaceStatus;
-
-#[derive(Debug, Deserialize)]
-pub struct CreateWorkspaceRequest {
-    pub workspace_admin_email: String,
-    pub workspace_name: String,
-    pub workspace_status: Option<WorkspaceStatus>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct UpdateWorkspaceRequest {
-    pub workspace_admin_email: String,
-    pub workspace_status: Option<WorkspaceStatus>,
-    pub mandatory_dimensions: Option<Vec<String>>,
-}
 
 #[derive(Deserialize, Debug)]
 pub struct WorkspaceListFilters {
