+++ crates/superposition_types/src/api/workspace.rs
@@ -1,0 +1,17 @@
+use serde::{Deserialize, Serialize};
+
+use crate::database::models::WorkspaceStatus;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct CreateWorkspaceRequest {
+    pub workspace_admin_email: String,
+    pub workspace_name: String,
+    pub workspace_status: Option<WorkspaceStatus>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct UpdateWorkspaceRequest {
+    pub workspace_admin_email: String,
+    pub workspace_status: Option<WorkspaceStatus>,
+    pub mandatory_dimensions: Option<Vec<String>>,
+}
